Improved code generator.
The new CodeGenerator class centralizes more of the context needed for generating code,
such as variable names and type information. The type system explicitly notes the
difference between the type information guaranteed to hold by the pattern match
and the information required for the precondition and target to be valid. In particular,
types which may be distinct in the source but must be equal in the precondition or target
will always generate a type equality clause.